Refactor osf unconfirmed account exception
