+++ cas-server-webapp/src/main/resources/messages.properties
@@ -144,11 +144,11 @@ screen.badworkstation.heading=You cannot login from this workstation.
 screen.badworkstation.message=Please contact <a style="white-space: nowrap" href="mailto:support@osf.io">support@osf.io</a> to regain access.
 
 # OSF Login Failure Pages
-screen.loginnotallowed.heading=Account not confirmed
-screen.loginnotallowed.message=The OSF account associated with the email has been registered but not confirmed.  Please check your email (and spam folder) or click the button below to resend your confirmation email.
+screen.accountnotconfirmed.osflogin.heading=Account not confirmed
+screen.accountnotconfirmed.osflogin.message=The OSF account associated with the email has been registered but not confirmed.  Please check your email (and spam folder) or click the button below to resend your confirmation email.
+screen.accountnotconfirmed.osflogin.button.resendConfirmation=Resend confirmation email
 screen.accountnotconfirmed.idplogin.heading=Account not confirmed
 screen.accountnotconfirmed.idplogin.message=The OSF account associated with the email has been registered but not confirmed. Our records show that this account was created via ORCiD login. Please check your email (and spam folder) for the confirmation link. If you believe this should not happen, please contact <a style="white-space: nowrap" href="mailto:support@osf.io">OSF Support</a>.
-screen.loginnotallowed.button.resendConfirmation=Resend confirmation email
 screen.accountdisabled.heading=Account disabled
 screen.accountdisabled.message=The OSF account associated with the email has been disabled.  Please contact <a style="white-space: nowrap" href="mailto:support@osf.io">OSF Support</a> to regain access.
 screen.shouldnothappen.heading=Account not active
